+++ crates/client/src/watch_main.rs
@@ -175,7 +175,7 @@ async fn main() -> Result<()> {
     tracing_subscriber::fmt()
         .with_env_filter(
             tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| "find_watch=info".into()),
+                .unwrap_or_else(|_| "warn,find_watch=info".into()),
         )
         .init();