+++ crates/server/src/main.rs
@@ -64,7 +64,7 @@ pub struct AppState {
 async fn main() -> Result<()> {
     tracing_subscriber::registry()
         .with(tracing_subscriber::EnvFilter::try_from_default_env()
-            .unwrap_or_else(|_| "find_server=info,tower_http=info".into()))
+            .unwrap_or_else(|_| "warn,find_server=info,tower_http=info".into()))
         .with(tracing_subscriber::fmt::layer())
         .init();